simplify transcription state management
